fix(db): make the migration chain apply from a clean checkout
`supabase start` failed on the very first migration, so no one could stand up a
dev or CI database, and no migration could be tested before reaching production.
Seven independent defects, none of which had ever executed:

- 001_create_schema.sql: stray duplicate that sorted ahead of the real
  001_mixhive_schema.sql (which 002 names as its predecessor). Contained 45
  MySQL-style INDEX clauses inside CREATE TABLE and a DELETE trigger whose WHEN
  referenced NEW. Deleted: production has mixes.dj_id from 001_mixhive_schema,
  not this file's mixes.user_id, confirming it never ran.
- 040 collision: 040_audio_jobs and 040_buzz_triggers_and_fix shared a prefix and
  schema_migrations keys on it. Renumbered audio_jobs to 022 — buzz keeps 040
  because 041 says "Run after 040_buzz_triggers_and_fix.sql", and audio_jobs
  only depends on public.mixes while needing to precede 042, which extends it.
- 000_enable_extensions.sql: new. 008 and 015 hard-fail without pg_cron by
  design; enabling it up front keeps their guards intact instead of neutering
  them. Idempotent, so a no-op where it was enabled by hand.
- 007: direct DML on storage.objects/buckets, which current Supabase forbids.
  Wrapped as best-effort cleanup.
- 009: joined follows.followed_id; the column is following_id (confirmed
  against production).
- 040_buzz: dropped handle_reply_notification while its trigger still depended
  on it. Added cascade; the trigger is recreated immediately below.
- 071: create-or-replace cannot change a function's return type. Drops existing
  overloads by signature first.
- 089: calculate_level takes integer but was passed sum()'s bigint.

All 111 migrations now apply. This also finally exercised migration 112's RLS,
which had been committed but unverified: host reads own private room (1),
participant read raises no 42P17 (1), stranger sees nothing (0), anonymous sees
nothing (0).

+++ supabase/migrations/089_xp_reputation.sql
@@ -70,7 +70,9 @@ create trigger tr_sync_xp_on_review
 update public.profiles p
 set 
   xp = sub.total_xp,
-  level = public.calculate_level(sub.total_xp),
+  -- sum(integer) yields bigint, and calculate_level takes integer; Postgres
+  -- will not implicitly narrow it for function resolution (42883).
+  level = public.calculate_level(sub.total_xp::int),
   reputation_score = sub.avg_rep
 from (
   select
